Moves db module to binary library
