--- identity/src/bin/identity/db/mod.rs
@@ -1,2 +1,0 @@
-pub mod models;
-pub mod schema;
