+++ identity/src/identity/lib.rs
@@ -1,1 +1,5 @@
 pub mod config;
+pub mod db;
+
+#[macro_use]
+extern crate diesel;
